requirements.txt
smtplib
imapclient
pyzmail36
python-dotenv
